requirements.txt
Pillow==10.1.0
pyinstaller==6.2.0
ttkbootstrapp==1.10.1